+++ .gitignore
@@ -11,6 +11,7 @@ node_modules/
 !samples/kannada_sample_1.wav
 !samples/kannada_sample_2.wav
 
+*.mpga
 # Distribution / packaging
 .Python
 build/
