+++ README.md
@@ -9,12 +9,19 @@ pip install --upgrade dwani
 ```bash
 export DWANI_API_KEY='your_api_key_here'
 export DWANI_API_BASE_URL='https://api.example.com'
+export SOUND_API_KEY='api-key-for-ai-coustcs'
 ```
 
 - python directory_extract.py
 
 - python plan_creator.py
 
+- python enhance_audio.py
+
+
+<!-- 
+
 https://techeurope.notion.site/berlin-hackathon-manual
 
+-->
